Refactor code directory structure
