+++ src/samplers/spatial/sampler_pseudo_nn.h
@@ -48,7 +48,7 @@
 #ifndef MUI_SAMPLER_PSEUDO_NN_H_
 #define MUI_SAMPLER_PSEUDO_NN_H_
 
-#include "../config.h"
+#include "../../config.h"
 #include "../sampler.h"
 
 namespace mui {
