+++ src/samplers/spatial/sampler_sph_quintic.h
@@ -51,8 +51,8 @@
 #ifndef MUI_SAMPLER_SPH_QUINTIC_H_
 #define MUI_SAMPLER_SPH_QUINTIC_H_
 
-#include "../util.h"
-#include "../config.h"
+#include "../../general/util.h"
+#include "../../config.h"
 #include "../sampler.h"
 
 namespace mui
