+++ src/samplers/temporal/temporal_sampler_null.h
@@ -48,8 +48,8 @@
 #ifndef MUI_TEMPORAL_SAMPLER_NULL_H_
 #define MUI_TEMPORAL_SAMPLER_NULL_H_
 
-#include "../util.h"
-#include "../config.h"
+#include "../../general/util.h"
+#include "../../config.h"
 
 namespace mui {
