+++ src/uniface.h
@@ -52,20 +52,20 @@
 #ifndef UNIFACE_H_
 #define UNIFACE_H_
 
-#include "util.h"
+#include "general/util.h"
 #include "communication/comm.h"
 #include "communication/comm_factory.h"
 #include "config.h"
-#include "dynstorage.h"
-#include "spatial_storage.h"
-#include "lib_dispatcher.h"
-#include "message.h"
-#include "reader_variable.h"
-#include "stream_vector.h"
-#include "stream_unordered.h"
-#include "stream_string.h"
+#include "storage/dynstorage.h"
+#include "storage/spatial_storage.h"
+#include "communication/lib_dispatcher.h"
+#include "communication/message/message.h"
+#include "communication/message/reader_variable.h"
+#include "storage/stream_vector.h"
+#include "storage/stream_unordered.h"
+#include "storage/stream_string.h"
 #include "storage/bin.h"
-#include "stream.h"
+#include "storage/stream.h"
 
 #ifdef PYTHON_BINDINGS
 #include <pybind11/pybind11.h>
